% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_number.R
\name{is_valid_mod11}
\alias{is_valid_mod11}
\alias{is_valid_mod11.default}
\alias{is_valid_mod11.numeric}
\alias{is_valid_mod11.character}
\title{Validate NHS number}
\usage{
is_valid_mod11(x, ...)

\method{is_valid_mod11}{default}(x, ...)

\method{is_valid_mod11}{numeric}(x, ...)

\method{is_valid_mod11}{character}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Not currently used.}
}
\value{
A Logical vector the same length as `x`.
}
\description{
Generic for testing whether input is a valid NHS number in terms of the
modulo 11 check. Methods are provided for character and numeric input.
Character input should only contain integer values (e.g. no hyphens or
spaces).
}
\note{
Validity is determined solely on whether of not the input passes the modulo
11 check. The number is not verified against numbers issued nor is any
additional checking of the first nine digits performed.
}
\examples{
dat <- c("5390502108", NA_character_, "2788584652", "3510670484")
is_valid_mod11(dat)

dat <- c(5390502108, NA_real_, 2788584652, 3510670484)
is_valid_mod11(dat)

}
\seealso{
https://www.datadictionary.nhs.uk/attributes/nhs_number.html for details on
the validation algorithm.
}
